cleanup

+++ infobase/lileg_db_bot.py
@@ -1,15 +1,11 @@
-import hashlib
 import logging.handlers
 import os
-from datetime import datetime, UTC
 from pathlib import Path
 
 import requests
 from dotenv import load_dotenv
 from langchain.docstore.document import Document
-from langchain_chroma import Chroma
 from langchain_community.document_loaders import PlaywrightURLLoader, PyPDFLoader
-from langdetect import detect
 from telegram import Update, InlineKeyboardMarkup, InlineKeyboardButton, Message, User
 from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes, MessageHandler, filters, \
     CallbackQueryHandler, CallbackContext
@@ -43,16 +39,22 @@ async def echo(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
 
 
 async def clear(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
-    collection_name = str(update.effective_user.id)
+    message = update.effective_message
 
-    vector_store = Chroma(
-        collection_name=collection_name,
-        embedding_function=EMBEDDINGS,
-        persist_directory=CHROMA_CLIENT_DIR,
+    logger.info("Removing existing messages from message id %s", message.message_id)
+
+    user_id = update.effective_user.id
+    chat_id = update.effective_chat.id
+
+    response = requests.post(
+        f'http://localhost:8000/users/{user_id}/chats/{chat_id}/forgetAll',
     )
 
-    logger.info("Removing existing messages from collection %s", collection_name)
-    vector_store.delete_collection()
+    if not response.ok:
+        raise Exception(
+            "Failed to forget all from message id %s! [%s]: %s",
+            message.message_id, response.status_code, response.text
+        )
 
     await update.effective_message.reply_text(
         f"Your data have been removed from the database!",
@@ -72,44 +74,21 @@ async def reply_message(message: Message, user: User):
     )
 
 
-def hash_text(text: str, algorithm="sha256", encoding="utf-8") -> str:
-    hasher = hashlib.new(algorithm)
-    hasher.update(text.encode())
-    return hasher.hexdigest()
-
-
-def ingest_documents(collection_name: str, documents: list[Document], message: Message):
-    if all(doc.page_content == "" for doc in documents):
-        logger.info("Failed to fetch documents from message id %s", update.effective_message.message_id)
-        raise Exception("All documents are empty")
-
-    vector_store = Chroma(
-        collection_name=collection_name,
-        embedding_function=EMBEDDINGS,
-        persist_directory=CHROMA_CLIENT_DIR,
-    )
-
-    logger.info("Calculating hash from message id %s", message.message_id)
-    message_text_hash = hash_text(message.text or message.document.file_name)
-
-    language = "unknown"
-    try:
-        logger.info("Detecting language from message id %s", message.message_id)
-        language = detect(message.text)
-    except Exception:
-        # langdetect.lang_detect_exception.LangDetectException: No features in text.
-        logger.warning("Failed to detect language from message id %s", message.message_id)
-
+def enrich_metadata(documents: list[Document], message: Message, metadata: dict[str, str]) -> list[Document]:
     logger.info("Enhancing document metadata from message id %s", message.message_id)
-    message_metadata = {
-        "hash": message_text_hash,
+    message_metadata = metadata | {
         "message_id": str(message.message_id),
         "message_date": str(message.date),
-        "ingested_on": str(datetime.now(UTC)),
-        "language": language,
     }
 
-    [x.metadata.update(message_metadata) for x in documents]
+    for document in documents:
+        document.metadata.update(message_metadata)
+
+        for k, v in document.metadata.items():
+            if not isinstance(v, str):
+                document.metadata[k] = str(v)
+
+    return documents
 
 
 async def ingest_text(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
@@ -120,17 +99,18 @@ async def ingest_text(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
     user_id = update.effective_user.id
     chat_id = update.effective_chat.id
 
+    documents = [
+        Document(page_content=message.text, metadata={"source": str(message.message_id)})
+    ]
+    documents = enrich_metadata(documents, message, {"source_type": "message"})
     response = requests.post(
-        f'http://localhost:8000/users/{user_id}/chats/{chat_id}/vectorize',
-        json=[{
-            "content": message.text,
-            "metadata": {"source": str(message.message_id)}
-        }]
+        f'http://localhost:8000/users/{user_id}/chats/{chat_id}/remember',
+        json=[{"content": x.page_content, "metadata": x.metadata} for x in documents]
     )
 
     if not response.ok:
         raise Exception(
-            "Failed to vectorize from message id %s! [%s]: %s",
+            "Failed to remember from message id %s! [%s]: %s",
             message.message_id, response.status_code, response.text
         )
 
@@ -140,29 +120,43 @@ async def ingest_text(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
 
 
 async def ingest_url(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
-    collection_name = str(update.effective_user.id)
+    message = update.effective_message
 
-    logger.info("Start indexing url from message id %s", update.effective_message.message_id)
+    logger.info("Start indexing url from message id %s", message.message_id)
 
     documents = await PlaywrightURLLoader(
-        urls=update.effective_message.text.splitlines(), remove_selectors=["header", "footer"]
+        urls=message.text.splitlines(), remove_selectors=["header", "footer"]
     ).aload()
 
-    ingest_documents(collection_name, documents, message=update.effective_message)
+    user_id = update.effective_user.id
+    chat_id = update.effective_chat.id
 
-    logger.info("Finish indexing url from message id %s", update.effective_message.message_id)
+    documents = enrich_metadata(documents, message, {"source_type": "url"})
+    response = requests.post(
+        f'http://localhost:8000/users/{user_id}/chats/{chat_id}/remember',
+        json=[{"content": x.page_content, "metadata": x.metadata} for x in documents]
+    )
 
-    await reply_message(update.effective_message, update.effective_user)
+    if not response.ok:
+        raise Exception(
+            "Failed to remember from message id %s! [%s]: %s",
+            message.message_id, response.status_code, response.text
+        )
+
+    logger.info("Finish indexing url from message id %s", message.message_id)
+
+    await reply_message(message, update.effective_user)
 
 
 async def ingest_file(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    collection_name = str(update.effective_user.id)
-    logger.info("Start indexing file from message id %s", update.effective_message.message_id)
+    message = update.effective_message
 
-    local_file_path = Path(".documents") / str(update.effective_user.id) / update.effective_message.document.file_name
+    logger.info("Start indexing file from message id %s", message.message_id)
+
+    local_file_path = Path(".documents") / str(update.effective_user.id) / message.document.file_name
     local_file_path.parent.mkdir(parents=True, exist_ok=True)
 
-    file = await context.bot.get_file(update.effective_message.document)
+    file = await context.bot.get_file(message.document)
     await file.download_to_drive(local_file_path)
 
     message = (update.message or update.edited_message)
@@ -170,9 +164,22 @@ async def ingest_file(update: Update, context: ContextTypes.DEFAULT_TYPE) -> Non
     logger.info("Loading content from pdf %s", local_file_path)
     documents = await PyPDFLoader(local_file_path).aload()
 
-    ingest_documents(collection_name, documents, message=update.effective_message)
+    user_id = update.effective_user.id
+    chat_id = update.effective_chat.id
 
-    logger.info("Finish indexing file from message id %s", update.effective_message.message_id)
+    documents = enrich_metadata(documents, message, {"source_type": "file"})
+    response = requests.post(
+        f'http://localhost:8000/users/{user_id}/chats/{chat_id}/remember',
+        json=[{"content": x.page_content, "metadata": x.metadata} for x in documents]
+    )
+
+    if not response.ok:
+        raise Exception(
+            "Failed to remember from message id %s! [%s]: %s",
+            message.message_id, response.status_code, response.text
+        )
+
+    logger.info("Finish indexing file from message id %s", message.message_id)
 
     await reply_message(message, update.effective_user)
 
@@ -182,17 +189,24 @@ async def keyboard_callback(update: Update, _: CallbackContext) -> None:
 
     logger.info("Received Keyboard callback %s", query.data)
 
-    collection_name = str(update.effective_user.id)
-    vector_store = Chroma(
-        collection_name=collection_name,
-        embedding_function=EMBEDDINGS,
-        persist_directory=CHROMA_CLIENT_DIR,
-    )
-
     command, message_id = query.data.split(":")
     if command == "delete":
-        logger.info("Deleting message %s from collection %s", message_id, collection_name)
-        vector_store.delete(where={"message_id": message_id})
+        logger.info("Deleting message %s", message_id)
+
+        user_id = update.effective_user.id
+        chat_id = update.effective_chat.id
+
+        response = requests.post(
+            f'http://localhost:8000/users/{user_id}/chats/{chat_id}/forget',
+            json={"filter": {"message_id": message_id}}
+        )
+
+        if not response.ok:
+            raise Exception(
+                "Failed to forget all from message id %s! [%s]: %s",
+                message_id, response.status_code, response.text
+            )
+
         await query.answer(f"Message {message_id} deleted!")
         await query.delete_message()
     else:
